agent/pyproject.toml
[project]
name = "a2ui-weather-app"
version = "0.1.0"
description = "Weather app using Google ADK with MCP tools, A2UI, and human-in-the-loop."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "google-adk>=1.8.0",
    "google-genai>=1.27.0",
    "python-dotenv>=1.1.0",
    "litellm",
    "a2ui",
    "jsonschema>=4.0.0",
    "mcp>=1.0.0",
    "fastmcp>=0.9.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
a2ui = { workspace = true }
